Adding Role Select
Added a role select page to display a message box asking if user is a physician. If user is not a physician then it redirects them to the added patient register page. As of now, role select page does not work, I cannot figure out how to add the missing assembly reference as it does not give me the option in the add reference menu.